used react useContext for avoiding prop drilling
