+++ src/components/ToDoCounter/ToDoCounter.js
@@ -1,6 +1,11 @@
+import React from "react";
 import "./ToDoCounter.css";
+import { ToDoContext } from "../ToDoContext/ToDoContext";
 
-function ToDoCounter({ total, completed }) {
+function ToDoCounter() {
+  const { completedToDos: completed, totalToDos: total } =
+    React.useContext(ToDoContext);
+
   if (total === 0 && completed === 0) {
     return <h1 className="ToDoCounter">You have no ToDos</h1>;
   } else if (total === completed) {
